+++ ontimize/layouts/app-layout/o-app-layout.component.ts
@@ -1,4 +1,4 @@
-import { Component, NgModule, ViewEncapsulation } from '@angular/core';
+import { Component, NgModule, ViewEncapsulation, EventEmitter } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { CommonModule } from '@angular/common';
 
@@ -22,7 +22,12 @@ export const DEFAULT_INPUTS_O_APP_LAYOUT = [
   'closedSidenavImg: closed-sidenav-image'
 ];
 
-export const DEFAULT_OUTPUTS_O_APP_LAYOUT: any[] = [];
+export const DEFAULT_OUTPUTS_O_APP_LAYOUT: any[] = [
+  'beforeOpenSidenav',
+  'afterOpenSidenav',
+  'beforeCloseSidenav',
+  'afterCloseSidenav'
+];
 
 export type OAppLayoutMode = 'mobile' | 'desktop';
 export type OSidenavMode = 'over' | 'push' | 'side';
@@ -60,6 +65,11 @@ export class OAppLayoutComponent {
   openedSidenavImg: string;
   closedSidenavImg: string;
 
+  beforeOpenSidenav: EventEmitter<boolean> = new EventEmitter<boolean>();
+  afterOpenSidenav: EventEmitter<boolean> = new EventEmitter<boolean>();
+  beforeCloseSidenav: EventEmitter<boolean> = new EventEmitter<boolean>();
+  afterCloseSidenav: EventEmitter<boolean> = new EventEmitter<boolean>();
+
   get showHeader(): boolean {
     return this._showHeader;
   }
@@ -95,6 +105,14 @@ export class OAppLayoutComponent {
     } else {
       console.error('Invalid `o-app-layout` sidenav-mode (' + val + ')');
     }
+  }
+
+  sidenavToggle(opened: boolean) {
+    opened ? this.beforeOpenSidenav.emit() : this.beforeCloseSidenav.emit();
+  }
+
+  afterToggle(opened: boolean) {
+    opened ? this.afterOpenSidenav.emit() : this.afterCloseSidenav.emit();
   }
 }